+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### v2.1.3
 - Fix auto-copy code
 - Fix detect paste data
+- Fix for notes folder not available
 
 ### v2.1.2
 - Fix for ToC